fix(setup): fix setup issue with pip install dependency on nltk
Pip was not able to find a version that satisfies the requirement 'nltk >= 3.0' although nltk3.0b1 is the last release.
This seems to be related to nltk having a  naming convention not supported by Pip (see http://stackoverflow.com/a/18319230) which seems to consider releases as pre-releases.
Including a valid release name for Nltk apparently makes it work.